feat(typage): harnais de parité dbt↔cœur + seam relevés en Int64 (#291)
Garde-fou de frontière (ADR-0035) : table de correspondance SQL↔Polars + helper
`ecarts_de_typage` réutilisable, câblés au seam releves()↔RelevéIndex. Le test lit
le type RÉELLEMENT émis par dbt (DESCRIBE), pas la sortie post-cast du loader —
sinon un re-typage loader blanchirait une dérive du modèle dbt (le mode d'échec
d'ADR-0034 : index ~1000× perdus en silence faute de garde-fou à cette frontière).

Types arrêtés (ADR-0034) : les index relevés deviennent Int64 (kWh entiers, le
type bigint que dbt émet nativement), de bout en bout sur ce seam :
- RelevéIndex / ChronologieReleves : index_*_kwh Float64 → Int64
- releves() : `SELECT *` strict (cast BIGINT→Float64 retiré ; transform=None)
- loaders /flux r151,r15 : col_cast_double → col_simple sur les index
- energie : la différence d'index est cast Float64 au boundary (TURPE/facturation
  inchangés ; le passage Int64 s'arrête au seam des relevés)

Cas négatif (test_parite_typage) : casser le type dbt OU le dtype Pandera rougit ;
un type SQL inconnu lève. Fixtures aval alignées sur Int64.

+++ electricore/core/models/chronologie_releves.py
@@ -59,14 +59,15 @@ class ChronologieReleves(pa.DataFrameModel):
     # 🏢 Référence calendrier distributeur (utile au contrôle des cadrans attendus)
     id_calendrier_distributeur: pl.Utf8 | None = pa.Field(nullable=True)
 
-    # ⚡ Index de compteurs (valeurs cumulées en kWh)
-    index_hp_kwh: pl.Float64 | None = pa.Field(nullable=True)
-    index_hc_kwh: pl.Float64 | None = pa.Field(nullable=True)
-    index_hch_kwh: pl.Float64 | None = pa.Field(nullable=True)
-    index_hph_kwh: pl.Float64 | None = pa.Field(nullable=True)
-    index_hpb_kwh: pl.Float64 | None = pa.Field(nullable=True)
-    index_hcb_kwh: pl.Float64 | None = pa.Field(nullable=True)
-    index_base_kwh: pl.Float64 | None = pa.Field(nullable=True)
+    # ⚡ Index de compteurs (valeurs cumulées en kWh entiers, ADR-0034 ; Int64 hérité de
+    # RelevéIndex — le loader ne re-caste plus, ADR-0035).
+    index_hp_kwh: pl.Int64 | None = pa.Field(nullable=True)
+    index_hc_kwh: pl.Int64 | None = pa.Field(nullable=True)
+    index_hch_kwh: pl.Int64 | None = pa.Field(nullable=True)
+    index_hph_kwh: pl.Int64 | None = pa.Field(nullable=True)
+    index_hpb_kwh: pl.Int64 | None = pa.Field(nullable=True)
+    index_hcb_kwh: pl.Int64 | None = pa.Field(nullable=True)
+    index_base_kwh: pl.Int64 | None = pa.Field(nullable=True)
 
     # 🚩 Flag de complétude : relevé attendu mais absent à cette date (relevés interrogés).
     releve_manquant: pl.Boolean | None = pa.Field(nullable=True)
